+++ Editor/Attributes/ConditionalField.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Linq;
-using System.Collections;
-using System.Reflection;
 using UnityEditor;
 using UnityEngine;
 
@@ -12,7 +10,7 @@ using CinderUtils.Extensions;
 namespace CinderUtils.Editor {
 
     [CustomPropertyDrawer(typeof(ConditionalFieldAttribute), true)]
-    public class ConditionalFieldAttributeDrawer : PropertyDrawer {
+    public class ConditionalFieldAttributeDrawer : CinderPropertyDrawer {
         private bool shouldBeDrawn = true;
 
         /// <param name="property">The serialized value of the "field" that is being drawn.</param>
@@ -27,14 +25,14 @@ namespace CinderUtils.Editor {
             // 2. If the field was found
             if (targetValue != null) {
                 // 3. Check if the field's value matches with one of the provided "compareValues".
-                shouldBeDrawn = condAttr.compareValues.Contains(targetValue.ToString());
+                shouldBeDrawn = condAttr.compareValues.Contains(targetValue);
 
                 // 4. Invert the logic if it was configured like so on the attribute.
                 shouldBeDrawn = condAttr.inverse ? !shouldBeDrawn : shouldBeDrawn;
             }
 
-            // Return the hight (-2 hides it)
-            if (!shouldBeDrawn) return -2;
+            // Return the height the drawer should have, negative one hides it.
+            if (!shouldBeDrawn) return -EditorGUIUtility.standardVerticalSpacing;
             else return EditorGUI.GetPropertyHeight(property, label);
         }
 
@@ -52,64 +50,6 @@ namespace CinderUtils.Editor {
             var child = GetValue(parent, fieldName);
 
             return child;
-        }
-
-
-        // Get the parent targetObject of some property
-        internal object GetParent(SerializedProperty property) {
-            // Sanitize the path for values inside of collections.
-            var path = property.propertyPath.Replace(".Array.data[", "[");
-
-            // Get the object to which the "field" (property) belongs
-            object obj = property.serializedObject.targetObject;
-            var elements = path.Split('.');
-
-            // Find parent recursivelly
-            foreach (var element in elements.Take(elements.Length - 1)) {
-                // Handle collection indexing
-                if (element.Contains("[")) {
-                    var elementName = element.Substring(0, element.IndexOf("["));
-                    var index = Convert.ToInt32(element.Substring(element.IndexOf("[")).Replace("[","").Replace("]",""));
-                    obj = GetValue(obj, elementName, index);
-                }
-                else {
-                    obj = GetValue(obj, element);
-                }
-            }
-
-            return obj;
-        }
-
-        // Get value of a field/property
-        //! Credit: https://discussions.unity.com/t/get-the-instance-the-serializedproperty-belongs-to-in-a-custompropertydrawer/66954/2
-        internal object GetValue(object source, string name) {
-            if (source == null) return null;
-
-            // Get the type 
-            var type = source.GetType();
-
-            // Get field's value
-            var field = type.GetField(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
-            if (field != null) {
-                return field.GetValue(source);
-            }
-            // Otherwise get property's value
-            else {
-                var property = type.GetProperty(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.IgnoreCase);
-                
-                if (property == null) return null;
-                else return property.GetValue(source, null);
-            }
-        }
-
-        // Get value of a field inside a collection
-        internal object GetValue(object source, string name, int index) {
-            var enumerable = GetValue(source, name) as IEnumerable;
-            var enm = enumerable.GetEnumerator();
-            while (index-- >= 0) { 
-                enm.MoveNext();
-            }
-            return enm.Current;
         }
     }