+++ Editor/Attributes/Disabled.cs
@@ -7,18 +7,22 @@ using CinderUtils.Attributes;
 
 namespace CinderUtils.Editor {
 
-    [CustomPropertyDrawer(typeof(DisabledAttribute))]
-    public class DisabledDrawer : PropertyDrawer {
+    [CustomPropertyDrawer(typeof(DisabledAttribute), true)]
+    public class DisabledDrawer : CinderPropertyDrawer {
 
         public override float GetPropertyHeight(SerializedProperty property, GUIContent label) {
             return EditorGUI.GetPropertyHeight(property, label, true);
         }
 
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label) {
-            GUI.enabled = false;
-            EditorGUI.PropertyField(position, property, label, true);
-            GUI.enabled = true;
+            bool prevStatus = GUI.enabled;
+            GUI.enabled = IsEnabled;
+
+            EditorGUI.PropertyField(position, property, label, false);
+            GUI.enabled = prevStatus;
         }
+
+        public virtual bool IsEnabled { get => false; }
     }
 
 }